Game logic is started
Bugs with the redraw and raise bet logic, will take a look at later. I don't think we had the redraw originally but it seemed to make sense as thats how the inspiration game. Later tonight I'm hoping to have the bugs worked out and hopefully the comparison logic for who wins the round.